jetty-based scanner manager was replaced by jdk scanner manager
